# tstime: hand-implement parseInt for specific needs of rfc3339 parsing.