@danderson 2020-04-14
Put `d` assignment back, and stole an optimization from `strconv.Atoi`. parseInt now runs 4.6x faster than before, 6.9ns/parse